backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pandas==2.2.0
openpyxl==3.1.2
chromadb==0.4.22
google-generativeai==0.3.2
python-dotenv==1.0.1 